requirements.txt
fastapi>=0.116.0
uvicorn>=0.35.0
vastai-sdk>=0.3.0
